backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
sqlmodel==0.0.16
pgvector==0.2.4

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# Geo
h3==4.1.0

# AI/ML Embeddings
fastembed==0.2.2
openai==1.12.0
numpy==1.26.4

# Redis and Task Queue
redis==5.0.1
arq==0.25.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilities
httpx==0.26.0
python-dotenv==1.0.1
